Merged PR 5: Incorporate automation toolkit as part of this repo
# TL;DR
--------

Include the `Dockerfile` for the required `tanzu-automation-toolkit` image in this repository.

# Details
---------

The cluster management pipeline depends on a fairly lightweight toolkit image with `kubectl`, `jq`, and the `tmc` CLI available. I was creating it independently but wanted to make sure if others wanted to use the pipeline then it was easy for them to build it.