Merged PR 10: Work with multiple clusters
TL;DR
------

Change from a parameterized pipeline that works on one cluster to a repository driven approach that works with multiple clusters.

Details
-------

Splits the single pipeline into a parent-child model with a top-level driver pipeline that invokes the single-cluster pipeline as a child. The top-level pipeline reads [another repository](https://dev.azure.com/cdantonio/cluster-management/_git/clusters) containing YAML descriptions of the clusters to provision on how they will be seen in TMC. It runs through the same steps as the last version, but uses variables instead of parameters so that it can be invoked one for each desired cluster.

This approach allows for self-service through pull requests to the cluster repository.  See the [README in the cluster repository](https://dev.azure.com/cdantonio/cluster-management/_git/clusters?path=%2FREADME.md&_a=preview) for more info.