Merged PR 14: Schedule nightly backup as part of configuration
TL;DR
------

Use the `velero` command to schedule a nightly backup

Details
-------

TMC data protection installs Velero to the cluster and provides a user interface for managing backups and restores. In the current version, it will show scheduled backups but doesn't allow you to configure them. Since schedule backups are a common need, add a scheduled backup as part of enabling data protection on the cluster.