Merged PR 15: Add missing display names
TL;DR
------

Polish up the pipelines by adding display names

Details
-------

Many of the tasks in the pipelines didn't have display names, leading to unhelpful names like "CmdLine" in the GUI. Updated the names for all stages, tasks, and jobs to make things look a little nicer when demoing.

Also encountered and error where branch names with a `/` in them caused the pipeline to fail and switch to using the `Build.SourceBranch` variable when calling the child pipeline.