Merged PR 16: Replaced hard coded values with variables
TL;DR
------

Some things were still hard coded specific to my universe, now they're not

Details
-------

The values of a few items describing the Azure context for the pipeline were still hard coded. This change fixes it so that the key value, trusted registry, and storage location for TMC data protection are all set as variables in the pipeline configuration.