Merged PR 20: Fix missed variable rename
Fix missed variable rename

TL;DR
-----

There was still a reference to `backupLocation`, now there's not

Details
-------

When I renamed `backupLocation` to `storageLocation` I missed one
of the references to it. It didn't show in testing because I was
working with pre-existing clusters and that code path only ran
for new ones.

The defect manifest today when I was creating two new clusters so
I jumped in and fixed it.